+++ outputs/x86_64-linux/src/olympians-generic.nix
@@ -30,22 +30,17 @@ let
         {
           modules.desktop.fonts.enable = true;
           modules.desktop.wayland.enable = true;
-          modules.desktop.gaming.enable = false;
           # modules.secrets.desktop.enable = true;
         }
       ];
-    home-modules =
-      (map mylib.relativeToRoot [
+    home-modules = (
+      map mylib.relativeToRoot [
         # common
         "home/linux/gui.nix"
         # host specific
         "hosts/olympians-${name}/home.nix"
-      ])
-      ++ [
-        {
-          modules.desktop.gaming.enable = false;
-        }
-      ];
+      ]
+    );
   };
 
   modules-niri = {